requirements.txt
Flask
Biopython